Connection should not know anything about its child AsyncConnection
Summary:
Some code was added to the destructor for the Connection class to handle cleanup of an async connection pool connection.  Unfortunately this code depended on information about the child class (AsyncConnection) which creates a kind of circular dependency.

Move this code into the destructor of the AsyncConnection class instead.  This requires that some of the private members of Connection become protected instead to make them visible to AsyncConnection.

Reviewed By: aditya-jalan, jupyung

Differential Revision: D30654951

fbshipit-source-id: 06cd9fac72aaf7b9217b227d239a13e9259b5d60