Pass shared_ptrs to operations in (semi)futures instead of raw pointers
Summary:
While working with the SyncMysqlClient I discovered that while the AsyncMysqlClient preserves shared_ptrs to all in-progress operations, the SyncMysqlClient doesn't (it is synchronous, so why would it need to).  However, due to lifetime issues, the shared_ptr goes out of scope before the operation completes.

Looking at the code, there was no obvious reason why we needed to use raw pointers to convert to [semi]futures.  Some simple changes to pass around the shared_ptr allows us to have much safer guarantees on the lifetime of the operations as they are processed.

A side benefit is shorter code as we only need one function per operation type instead of two.  We can also remove some instances of passing shared_ptrs by reference (Ugh!)

Reviewed By: abal147

Differential Revision: D33485448

fbshipit-source-id: 9a4633112a2dc0e01cf88a0d59cfcde948c6a292