Add `get(col, default)`
Summary:
The API provided a `get()` that would throw an exception on NULL and a `getWithDefault()` where the caller would supply a default value that would be returned when the column was NULL.

Needing two functions for this is extra friction that we don't need.  A `get()` function with two parameters is much less friction (as well as a being much shorter).  Implement it and redirect the existing `getWithDefault` to use the new function instead.

Reviewed By: abal147

Differential Revision: D33569673

fbshipit-source-id: 1b129c15364a51fee480d780fc7d7e5e3082de23