Support local files in the sync client
Summary:
The async client doesn't support local files (used for LOAD DATA LOCAL INFILE) as the response from the server is telling the client to start sending the file and the code has never been written to handle this correctly in a background thread.

However, the sync client can support it.  Add the code to query the client for whether it can be supported when initializing the connection.

Reviewed By: abal147

Differential Revision: D33781238

fbshipit-source-id: da8c2ee3f635682a82ef4fa5144ccbbd69ed0302