Support a non-throwing fieldIndex() function
Summary: Customers need to be able to find out if a name is a valid column in a result set.  The current method to do this throws an exception if the name is not a valid column name.  This makes extra work for clients that don't have complete information on the queries they are processing.

Reviewed By: aditya-jalan

Differential Revision: D33821803

fbshipit-source-id: 564cd89b546846458a20a561e0fcaaf351c22a1a