Update connection pool whenever settings change
Summary:
We would like to be able to update the connection pool settings whenever the config changes, but the connection pool code does not allow this.  Instead we will create a new pool and let the old one expire.

In order to handle this correctly I added a test to verify that we can update the connection pool while it is being used.

Reviewed By: abal147

Differential Revision: D34831727

fbshipit-source-id: 1d3e83c0e23a12ca9b6243d67f0d401b0ee669a4