Remove boolean as an option for QueryArguments
Summary:
While working on the Query class I discovered that there is a discrepancy in the handling of boolean values.  When assigned directly (`QueryArgument(bool)`) the boolean value is converted to an int64_t to store it.  When assigned from a folly::dynamic which has a bool type, the boolean is stored as a `bool` and then converted to an integer during rendering.

This doesn't make much sense, so in this diff I am removing bool as a special case and always converting it to int64_t internally.

Reviewed By: fadimounir

Differential Revision: D36325182

fbshipit-source-id: a8b8377455810960f5332c546514555f1f99ec8f