Add timeout information to error
Summary:
We recently had a case where we saw connections timing out with very small times (less than a millisecond).  This was confusing, but adding the timeout value that the code is using into the error message should help in the future.  Likely it will be the same as the delta already there, but it will make it clear what the user set.

In addition, I cleaned up some of the code generating the error message to allow me to share code better and the the timeout part is a single function used by connections, queries, and the connection pool.

Reviewed By: jasson15

Differential Revision: D36523848

fbshipit-source-id: 560e6179605d74f72ec958018ebe34a2a3bcb03e