Operations derived from ConnectOperation must not update TLS session resumption data in SSLOptionsProvider
Summary: Operation classes derived from ConnectOperation must not update TLS session resumption data in SSLOptionsProvider. There is a chance that the derived operation object is not directly involved in the creation of a new connection, and merely reuses a connection which was created earlier by another ConnectOperation. In this case the SSLOptionsProvider object associated with this operation can get updated with a TLS session resumption data created from a different client cert.

Reviewed By: abal147

Differential Revision: D36611060

fbshipit-source-id: 153737187edde777c1c980779066e2d3de7fe452