always send reset_connection when a conn is destroyed, without blocking
Summary:
This diff takes the opposite approach from what I recently did in the async conn pool, in which I tried to always send reset_connection in a delayed manner just before a conn is reused. The previous approach is beneficial in (1) avoiding wait operation in the conn destructor, and (2) sending reset_connection only when the conn is reused. But its problem is that (1) reset_connection can be sent to mysql gateway too late, which means that backend conn might be tied to its frontend conn longer than before, and (2) pooled conn operation can take longer than before because it always includes sending reset_connection.

So in this diff, I'd like to explore its opposite direction, always sending reset_connection earlier when a conn is destroyed, but this time without waiting for reset_connection to be completed. By not waiting for reset_connection to be completed, conn destructor can get faster. Also, by always sending reset_connection earlier just before the conn is released back to the conn pool, mysql gateway can unpin its backend conn from the frontend conn, making it ready to be reused. One problem, though, is that since conn destructor doesn't wait reset_connection any more, if the same conn is requested immediately after the previous conn is destroyed, it may end up opening a new connection, not reusing existing connection. This problem is reproducible in AsyncPoolTest, where I had to add explicit sleep function before the attempt to open a new conn, otherwise previous conn cannot be guaranteed to be reused (because reset_connection may not be completed yet).

So there's pros and cons when it comes to handling reset_connection eagerly or lazily, and I am leaning more towards the eager one, because with enough number of idle connections in the pool, the aforementioned conn miss issue will become almost no issue.

Reviewed By: jkedgar

Differential Revision: D38922901

fbshipit-source-id: c818aeceb92a57212c0f5334e566135fe98c2211