Refactor pooling to eventually support synchronous connection pooling
Summary:
In order to eventually support synchronous connection pooling the pool code needs to be refactored considerably.  This diff pulls the shared code into ConnectionPoolBase and ConnectionPool classes, leaving just the async specific code in the AsyncConnectionPool.

The Synchronous code that will take advantage of this will come in a later diff.

Reviewed By: jupyung

Differential Revision: D38911837

fbshipit-source-id: 6566d815269b083d16eb63f971ba471794b5feb0