Support both sync and async in pool storage
Summary: The existing pool storage code took advantage of the fact that the async code ran everything in a single thread - the mysql thread.  This won't work for the synchronous version.  In order to solve that refactor a bit so that the top level class either holds the storage directly (async version) or in a folly::Synchronized (sync version) and calls into it with appropriate accesses.  This will allow the same code to be used in both cases without paying an extra penalty in the async version.

Reviewed By: fadimounir

Differential Revision: D38913179

fbshipit-source-id: c249198927c50e17a9a772f7142b75b014e0fcc8