Pass pool options and markdown config by value
Summary: The existing functions to create clients and pools pass in PoolOptions and MarkdownConfigs by reference, but they always make a copy.  It would be more efficient to push the copy up a level.  This means that if the client creates a temporary (like the defualts do), we can use the temporaray directly instead of having to make a copy of it.

Reviewed By: fadimounir

Differential Revision: D38911916

fbshipit-source-id: d0da1e2396b3b0cfec86f0af4fae442a5d310fe2