Fix unused variable
Summary:
In recent changes for synchronous connection pool, the connection pool code migrated from the .cpp file to the .h file.  In the process existing code that had an unused variable could cause some builds to fail if certain warnings (like unused variables) were treated as errors.

Fix this by including the exception's error message in the logged output.  This is not really needed but it doesn't hurt anything.

Reviewed By: abal147

Differential Revision: D39318994

fbshipit-source-id: 6b03479c9da35d3dcbfa853d52b7d007211da405