Fix unused variable
Summary: For an overridden function in a derived class where the base version had an optional argument, I thought I needed to match it even though the base class doesn't access the variable.  This caused warnings which are treated as errors in some build environments.  It turns out I was wrong and the signatures don't have to match on the optional part.  Remove the variable name to avoid a warning.

Reviewed By: ibrahimjirdeh

Differential Revision: D39398985

fbshipit-source-id: d13105bde5a8095d92e93e7d4b81f3fa999020d1