Enforce a max connection timeout
Summary:
Enforce a maximum connection timeout without impacting pooled
operations.

This allows pooled conns to set the connect timeout and
influence how long to remain queued for available conns,
while not choking up the thread with connects that are
unlikely to succeed.

Add a GFLAG as a backdoor for those who insist on doing it
the wrong way

Reviewed By: jupyung

Differential Revision: D45589828

fbshipit-source-id: 3d4642daa530848ea8ff75d822b8b7b6251a195b