Add a lambda to the throttling callback to allow it to set the error information for throttling.
Summary:
The initial throttling callback was a simple yes/no setting, but if throttling was needed there was no way to set specific error information for the connection operation being throttled.  And unfortunately the current throttling error we use is not visible in Squangle.

To solve this, add a function that can be passed around that gets executed when throttling happens that can set the error number and error information.

Reviewed By: jupyung, abal147

Differential Revision: D44633039

fbshipit-source-id: da6cb037568a769a212028dcfde7f44eb63ea0d8