Switch attributes maps to F14 maps
Summary:
By default `std::unordered_map`s are not transparent.  That means that if you have a `std::string` as a key you _have_ to pass a `std::string` to `.find()`.  A transparent map would allow you to pass a `std::string_view` or `const char *`.  This means that we have to construct a `std::string` just to do lookups.

I tried to update the existing `std::unordered_map` to a version with a non-default hasher that would make them transparent, but it caused problems because we need to deserialize these maps for context related code and there is support for standard `std::unordered_map`s (https://fburl.com/code/axfjlv8v), but not for ones that have non-standard hashers.

On the other hand, Folly::F14 maps default to transparent.  Switch over all the attribute related maps from `std::unordered_map` to `F14::NodeMap`.

The only drawback to this change is that we can no longer use `.merge()`.  In `std::unordered_map`, merge allows us to move full nodes (both key and value) from one map to another without reallocating them.  Eventually it would be nice to see `merge` supported in F14, but it isn't critical.

Reviewed By: fadimounir

Differential Revision: D47197821

fbshipit-source-id: 6e61d76f986dab4a454538e75d8d991a69e71cf6