Modernize ConnectionKey
Summary:
Update ConnectionKey.  It was a class but had all members public.  Make them private and provide accessors.

This was driven by the fact that in order to protect the members they were marked `const`.  A `const` member cannot be moved - it has to be copied even when we are moving the entire ConnectionKey.  And we move ConnectionKeys often.  By making these non-`const` we can allow the move constructor to move the strings and not copy them

Reviewed By: aditya-jalan

Differential Revision: D47198978

fbshipit-source-id: 36ad7fa2dfbe703104a1315c8f8b6235e0fb8f31