Don't reuse pooled connections that change database
Summary:
When someone uses pooled connections in C++ and then issues a "USE <db>" command the connection pool was returning the connection to the pool without being aware that the connection was for a different database than expected.  This could cause errors or incorrect data to be returned.

Theoretically we could keep track of the database and move the connection to a different key in the pool but this is complicated.  Since we don't actually expect this to occur often (no one has complained about this - it was first noticed when using shadowdb replayer) a simpler solution is just not to reuse connections that have issued "USE" queries.

This code causes any connection that that changes the current database (even if it gets changed to the original database) to not be reused by the connection pool.

Reviewed By: fadimounir

Differential Revision: D50372225

fbshipit-source-id: a414b7f6203a0a7420da89ef42d39ce7562cbd51