On QueryException provide the ability to get the connection back out for reuse
Summary:
The existing Squangle implementation does not give a connection back to the client when an exception occurs.  This means that on a simple error like a syntax error or --read-only error the client has to generate a new connection to do more work.

We should provide the option of continuing to use the connection (provided the error wasn't a lost connection).

Differential Revision: D52610876

fbshipit-source-id: 8485b209d2d0c4b3edd784492872bca164d4be18