Back out "On QueryException provide the ability to get the connection back out for reuse"
Summary:
After playing around with this for a bit I decided it would be safer NOT to have this ability.

The main reason is that the connection is not for sure in a known state, particularly on a timeout. which seems like the most likely reason to use this feature.  If using the sync client it would probably be fine - the MySQL server itself enforces the timeout so it would no longer be using the connection.  However, with the async client the client code itself enforces the timeout and it is possible that the query is still executing on the server when the connection is returned to the customer.  If they then try to execute another query we would get unexpected/undefined behavior.

Reviewed By: aditya-jalan

Differential Revision: D57304152

fbshipit-source-id: 7314f25fa8d3d3f8790146d99e3e22ce6119f768