Avoid copying query text every time
Summary:
For logging purposes we are copying the query string into the   `QueryLoggingData` structure on _every_ query even though we only log a very small fraction of all queries.

If we switch to storing the query data in a shared_ptr we can copy the shared_ptr into the logging data instead and avoid copying the query string (which might be very long).

Differential Revision: D60054414

fbshipit-source-id: bb11b174e2cc1f6e3b61f0f33b73de8f83eed3ce