Stop using raw pointer to ConnectionKey
Summary:
We should avoid using raw pointers when we can.  Squangle passed around a raw pointer to the ConnectionKey frequently even though a ConnectionKey was always required.  Change this to pass around a reference instead.

Note that as part of cleaning this up the `SquangleLoggingData` object was changed from a struct to a class with getters instead of direct access to memebers.

Reviewed By: aditya-jalan

Differential Revision: D49613234

fbshipit-source-id: b7711ac9293434f76aaa40313491d2678178cb5e