Split Operation.[h|cpp]
Summary: The Operation.h and Operation.cpp files have lots of code for all the different derived classes.  Break most of the derived classes into their own files to make reading the code easier.  This is part of an effort to cleanup the Squangle code while refactoring for future development.

Differential Revision: D52370883

fbshipit-source-id: e735f4edf8882e6a76c492d90676818a971e9974