Simplify handling of events and timeouts
Summary:
The ConnectionSocketHandler class is a helper class that kind of abstracted the handling of event base events and timeouts.  However, the class needed the operation as a pointer to manage anything.  It would make more sense to derive Operation directly from the two classes (EventHandler and AsyncTimeout) directly.  This avoids extra indirection as well as a class of errors if the `op` (Operation) pointer never got set.

This also ends up simplifying a bunch of stuff.

Reviewed By: fadimounir, aditya-jalan

Differential Revision: D52306543

fbshipit-source-id: 1c52a764975af408404d7430da7b38d1e40a5597