Split Operation.[h|cpp] (try 2)
Summary: The Operation.h and Operation.cpp files have lots of code for all the different derived classes.  Break most of the derived classes into their own files to make reading the code easier.  This is part of an effort to cleanup the Squangle code while refactoring for future development.

Reviewed By: aditya-jalan

Differential Revision: D60970119

fbshipit-source-id: f24aa2c6a447b6d425346081b92f6f4e81dfb093