Split Operation into two classes
Summary:
The Operation class combined both the user facing API (`run()`, `wait()`, etc.) with the implementation details of the MySQL protocol.  In order to add another protocol option we need to abstract the implementation details away.

In order to achieve this the Operation class is split into an Operation and an OperationImpl.

Reviewed By: aditya-jalan

Differential Revision: D53026126

fbshipit-source-id: b9423480f0f7f3a2be36d82fdc4abf678be899e1