Refactor ConnectionKey
Summary: To support more than one protocol in Squangle we can no longer have a single ConnectionKey class.  Instead we need ConnectionKey to be a base class and then have a MysqlConnectionKey that derives from it.  Other protocols in the future would then have their own derived class.  This diff is also changing all public APIs that deal with ConnectionKeys to use shared_ptrs instead of direct access to the ConnectionKey.

Differential Revision: D54641796

fbshipit-source-id: c3cdd8ca94628dd28748fb02c5fed3726696f1d8