Don't assert that we are in completed state for `elapsed()`
Summary:
The original code asserted that we were in a completed state if `elapsed()` was called.  However, in the refactoring that was done we can now call this in cases where we are not in a completed state.  In particular if we have retries we don't mark the state as complete, but we can still log and the logging needs the elapsed time.

Remove the check.  It isn't necessary.

Differential Revision: D61628805

fbshipit-source-id: 2d2b583c24b771322c0597a58f0590f16d9ab08f