Correctly handle multi-queries passed into single query API
Summary:
Prior the refactoring work on Squangle calling the single query API with a multi-query generated some sort of exception, but during the refactor this instead hit a failed CHECK() and crashed.  We had no test in our own test suite for this but a customer ran into it (https://fb.workplace.com/groups/mysql.users/posts/26881187268169867/).

Here I've modified the code to detect the situation and return an appropriate exception for the issue.  In addition I've added a test to check for this in the MySQL C++ Client to prevent regressions in the future.

Reviewed By: fadimounir

Differential Revision: D62663718

fbshipit-source-id: 72aa3c4381c7f927b5a931a01685954ec421e3df