mysql_store_result and mysql_use_result can return nullptr
Summary:
These two functions can return null and we are just putting this into a MysqlResult which assumes that it will not be null.  We should not be building a MysqlResult under this scenario.

Something in the recent refactor seems to have changed what happens around this code and TAO is seeing an occasional crash as a result (S448897).  This may resolve the issue.

Reviewed By: fuu

Differential Revision: D63276382

fbshipit-source-id: 365896635d9de45c434fe4df290b6ce53afb76cd