Remove temporary fix
Summary: While working on S448897 I generated a quick diff to try to handle an error condition.  The check should not be necessary as we should never get to these lines of code with a null `res_`.  Later I found the real problem and fixed it in D63276382 which has landed.  I can now revert this hack.

Differential Revision: D63646255

fbshipit-source-id: e233ddd9cdc9315778bbbd1e28d5f35fe4b69ca7