Introduce StorageRow as a better way to store MySQL row/column data
Summary:
This diff introduces a new class called StorageRow which is designed to store MySQL row data in a space-efficient manner. The class is intended to work when the protocol returns data in string format or in native format, and it stores data in a way that allows for efficient retrieval and manipulation.

The original implementation used a minimum of 8 bytes plus a bit for each column.  The bit was for indicating NULL values and the 8 bytes was an offset in a data buffer where the actual data was written.  This resulted in `8 bytes + 1 bit + actual data contents` data needed.

In this version we have 4 bytes plus actual data written, where the actual data written can be optimized (i.e. only write 1 byte for integers near 0).  This is accomplished by splitting handling of small things (anything less than 4K) directly while offloading any column larger than 4K in separate storage.  The 4 bytes are used as 3 bits holding the type of data and 29 bits holding an offset into the data buffer.  29 bits can reference 512M and since only columns smaller than 4K are stored in the referenced space, we could theoretically handle 131,000 columns.  MySQL has a maximum number of columns per table of 4096, so this would support a result set with 32 tables joined with 4096 columns per table which should be way more than needed.

The final result is that for NULLs and booleans, only 4 bytes are needed.  For integers, we have 4 bytes of overhead plus either 1, 2, 4, or 8 bytes of data (depending on how close to 0 the value is).  Doubles take 4 bytes + 8 bytes.  Strings shorter than 4K take 4 bytes plus the size of the string.  Long strings are stored in a separate storage but take 4 bytes plus a variable integer size in main storage data section.

Differential Revision: D63793585

fbshipit-source-id: 7de28c2066bd831b0da32145fe0a1d6024183621