Correctly handle returning data when NULL
Summary:
When refactoring the Squangle code to handle data arriving in native format, I accidentally broke what happens when reading data that is null.  Originally it would throw a `folly::ConversionError` (except for strings which would be empty), but I accidentally generated a failure using DCHECK.

Revert to the correct behavior and add a test for the future.

Differential Revision: D65502033

fbshipit-source-id: 96b79ebd535d92c6d53b259cbec2680a7c85dcbc