Fix crash in Thrift protocol
Summary:
In D65959868 I added the ability to get db_version when going straight to MySQL.

In D66184293 I fixed a bug in that implementation when a failed connection attempts to log this information when using the MySQL protocol.

Unfortunately, the Thrift protocol has the same problem (here: https://fburl.com/code/erpqxl3l).

Thinking it through more, it makes more sense for the `serverInfo()` function to return an empty string when there is no valid connection than to throw an exception, so I am making that change and undoing part of the fix for D66184293 since the `serverInfo()` function is now safe to call no matter the state of the connection.

Differential Revision: D66253298

fbshipit-source-id: 15e959c34c58295d81715c8cfd58da947227da17