Allow fetch operations to free RowStream objects before connection is closed
Summary: When using a streaming version, it is possible for the fetch operation to be cancelled after receiving some rows.  In this case the RowStream objects need to be released before the connection can be closed.  This change allows this to happen by making the `cancel()` function a virtual function and overriding it in the fetch operation to do the free before continuing with the normal cancel.

Reviewed By: fadimounir

Differential Revision: D67400934

fbshipit-source-id: 3272ea77b52446716cfbbd1009c97d76a0d33f60