Fix bug introduced in D67400934
Summary:
In D67400934 the RowStream object (stored in `current_row_stream_`) would be freed on a cancel.  However, there was at least one code path where we could attempt to access it again in this scenario.  This broke a handful of tests.

Fix the code to correctly handle this scenario.

Differential Revision: D67909302

fbshipit-source-id: cfad60ec70521378b73bd4e14a9e20324e30f0f2