Use weak_ptr for holding operation during certificate validation (take 2)
Summary:
Customers have run into an occasional crash where the Operation appears to have been deallocated when certificate validation occurs (P1625691373).

My current theory is that the connect operation is being cancelled or timed out and freeing the operation even though it is still processing and OpenSSL calls the callback and attempts to use the memory from the operation.

In this diff I attempt to hold a weak_ptr to the operation for use in the certificate verification callback.  When we lock the weak_ptr to generate a shared_ptr, if it fails because the operation is gone we will just return an error and not attempt to access the memory.

Note: this is a second attempt.  The first (D66730727) caused some problems in HHVM that I had to resolve.

Reviewed By: mdko

Differential Revision: D67996104

fbshipit-source-id: bc61c688aff27e1421342bcc562b0b694593bbfe