Add support for tuples
Summary:
Add support for tuples to easyquery lib. Stuff like:
  WHERE (a, b) = (1, 2)
or
  WHERE (a, b) < (1, 2)
or
  WHERE (a, b) IN ((1, 2), (3, 4))
It is possible this is already doable in the existing lib but I couldn't figure out how in a nice way. Actually I still couldn't figure out how to do it in a nice way, but I wrapped up the complexity so lib users don't have to deal with it (they make a Tuple object with a list of values, which I convert to a Query object to embed in the larger query).

Differential Revision: D69356885

fbshipit-source-id: 212e7f481bb2cada9a12dfd2d3081a26812473ad