Remove `mustSucceed()` function and RequiredOperationFailedException
Summary:
The `mustSucceed()` function is dangerous.  It was intended as a shortcut during testing, but there are other ways to do that.  Remove it so that it won't be used in production code accidentally.

The RequiredOperationFailedException was generated to be thrown by this function so remove it (at least as a globally available exception) as well.

Reviewed By: tensor-flower

Differential Revision: D72977621

fbshipit-source-id: 8162592d8761acbbb7391d167f19f2bc043ae6b1